Buscador de ciudades pero tdvia no trae clima

+++ app/src/main/java/com/example/segundoparcial/presentacion/ciudades/CiudadesIntencion.kt
@@ -1,9 +1,7 @@
 package com.example.segundoparcial.presentacion.ciudades
 
-import com.example.segundoparcial.presentacion.clima.ClimaIntencion
-
 sealed class CiudadesIntencion {
 
     data class Buscar( val nombre:String ) : CiudadesIntencion()
-    data class Seleccionar(val indice: Int) : CiudadesIntencion()//Hay que modif
+    data class Seleccionar(val indice: Double) : CiudadesIntencion()//Hay que modif
 }
